# Conflicts:
#	src/Themes/KrakeAdmin/Styles/krakeadmin.css
#	src/Themes/KrakeAdmin/Styles/krakeadmin.min.css

+++ src/Modules/Contrib.Widgets/Controllers/AdminController.cs
@@ -13,6 +13,7 @@ using Orchard.UI.Notify;
 using Orchard.Widgets.Models;
 using Orchard.Widgets.Services;
 using System;
+using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Web.Mvc;
@@ -52,6 +53,10 @@ namespace Contrib.Widgets.Controllers {
                 _services.TransactionManager.Cancel();
                 return View(model);
             }
+            // Up until here this is replicating the behaviour of the Contents AdminController.
+            // However here we take a different approach and we don't publish, because we know that 
+            // the contentItem isn't "finished", since we got here not by hittng "Save" or "Publish",
+            // but by attempting to add a widget.
 
             _services.Notifier.Information(string.IsNullOrWhiteSpace(contentItem.TypeDefinition.DisplayName)
                 ? T("Your content has been created.")
@@ -61,17 +66,23 @@ namespace Contrib.Widgets.Controllers {
         }
 
         public ActionResult ListWidgets(int hostId, string zone) {
-            var widgetTypes = _widgetsService.GetWidgetTypeNames().OrderBy(x => x).ToList();
+            IEnumerable<string> widgetTypes = _widgetsService.GetWidgetTypeNames().OrderBy(x => x);
 
             var widgetsContainerPart = _contentManager.Get(hostId, VersionOptions.Latest).As<WidgetsContainerPart>();
             var settings = widgetsContainerPart.Settings.GetModel<WidgetsContainerSettings>();
             if (!settings.UseHierarchicalAssociation) {
                 if (!string.IsNullOrWhiteSpace(settings.AllowedWidgets))
-                    widgetTypes = widgetTypes.Where(x => settings.AllowedWidgets.Split(',').Contains(x)).ToList();
+                    widgetTypes = widgetTypes.Where(x => settings.AllowedWidgets.Split(',').Contains(x));
             } else {
                 var allowedWidgetsForZone = settings.HierarchicalAssociation.Where(x => x.ZoneName == zone).SelectMany(z=>z.Widgets.Select(w=>w.WidgetType));
-                widgetTypes = widgetTypes.Where(x => allowedWidgetsForZone.Contains(x)|| allowedWidgetsForZone.Contains("All")).ToList();
+                widgetTypes = widgetTypes.Where(x => allowedWidgetsForZone.Contains(x)|| allowedWidgetsForZone.Contains("All"));
             }
+            // filter widgetTypes by permission
+            widgetTypes = widgetTypes
+                .Where(wt => _services.Authorizer.Authorize(
+                    Orchard.Core.Contents.Permissions.CreateContent,
+                    _contentManager.New(wt)));
+
             var viewModel = _services.New.ViewModel()
                 .WidgetTypes(widgetTypes)
                 .HostId(hostId)
@@ -87,6 +98,12 @@ namespace Contrib.Widgets.Controllers {
             var widgetPart = _services.ContentManager.New<WidgetPart>(widgetType);
             if (widgetPart == null)
                 return HttpNotFound();
+
+            if (!_services.Authorizer.Authorize(
+                Orchard.Core.Contents.Permissions.CreateContent, 
+                widgetPart, 
+                T("Cannot create widget")))
+                return new HttpUnauthorizedResult();
 
             try {
                 var widgetPosition = _widgetManager.GetWidgets(hostId).Count(widget => widget.Zone == widgetPart.Zone) + 1;
@@ -116,6 +133,12 @@ namespace Contrib.Widgets.Controllers {
         [HttpPost, ActionName("AddWidget")]
         [Orchard.Mvc.FormValueRequired("submit.Publish")]
         public ActionResult AddWidgetPublishPOST(string widgetType, int hostId) {
+
+            var dummyContent = _contentManager.New(widgetType);
+            if (!_services.Authorizer.Authorize(
+                Orchard.Core.Contents.Permissions.PublishContent, dummyContent, T("Couldn't create and publish widget")))
+                return new HttpUnauthorizedResult();
+
             return AddWidgetPost(widgetType, hostId, contentItem => _services.ContentManager.Publish(contentItem));
         }
 
@@ -125,12 +148,24 @@ namespace Contrib.Widgets.Controllers {
             if (!IsAuthorizedToManageWidgets())
                 return new HttpUnauthorizedResult();
 
+            // Check that the user has permission to create the widget before doing anything
+            var dummyWidget = _services.ContentManager.New<WidgetPart>(widgetType);
+            if (dummyWidget == null)
+                return HttpNotFound();
+            if (!_services.Authorizer.Authorize(
+                Orchard.Core.Contents.Permissions.EditContent,
+                dummyWidget,
+                T("Cannot create widget")))
+                return new HttpUnauthorizedResult();
+
             var layer = _widgetsService.GetLayers().First();
             var widgetPart = _widgetsService.CreateWidget(layer.Id, widgetType, "", "", "");
-            if (widgetPart == null)
+            if (widgetPart == null) {
                 return HttpNotFound();
-            else if(!widgetPart.ContentItem.TypeDefinition.Settings.GetModel<ContentTypeSettings>().Draftable &&  !widgetPart.ContentItem.Has<IPublishingControlAspect>())
+            } else if (!widgetPart.ContentItem.TypeDefinition.Settings.GetModel<ContentTypeSettings>().Draftable
+                  && !widgetPart.ContentItem.Has<IPublishingControlAspect>()) {
                 _contentManager.Publish(widgetPart.ContentItem);
+            }
 
             var contentItem = _services.ContentManager.Get(hostId, VersionOptions.Latest);
 
@@ -170,6 +205,9 @@ namespace Contrib.Widgets.Controllers {
                 _services.Notifier.Error(T("Widget not found: {0}", id));
                 return Redirect(returnUrl);
             }
+            if (!_services.Authorizer.Authorize(
+                Orchard.Core.Contents.Permissions.EditContent, widgetPart, T("Cannot edit widget")))
+                return new HttpUnauthorizedResult();
             try {
                 var model = _services.ContentManager.BuildEditor(widgetPart).HostId(hostId);
                 return View(model);
@@ -183,6 +221,15 @@ namespace Contrib.Widgets.Controllers {
         [HttpPost, ActionName("EditWidget")]
         [Orchard.Mvc.FormValueRequired("submit.Publish")]
         public ActionResult EditWidgetPublishPOST( int hostId, int id) {
+            var content = _contentManager.Get(id, VersionOptions.Latest);
+
+            if (content == null)
+                return HttpNotFound();
+
+            if (!_services.Authorizer.Authorize(
+                Orchard.Core.Contents.Permissions.PublishContent, content, T("Couldn't publish widget")))
+                return new HttpUnauthorizedResult();
+
             return EditWidgetPost(hostId, id, contentItem => {
                 _services.ContentManager.Publish(contentItem);
             });
@@ -210,6 +257,10 @@ namespace Contrib.Widgets.Controllers {
 
             if (widgetPart == null)
                 return HttpNotFound();
+
+            if (!_services.Authorizer.Authorize(
+                Orchard.Core.Contents.Permissions.EditContent, widgetPart, T("Cannot edit widget")))
+                return new HttpUnauthorizedResult();
 
             try {
                 var model = _services.ContentManager.UpdateEditor(widgetPart, this).HostId(hostId);
@@ -249,6 +300,9 @@ namespace Contrib.Widgets.Controllers {
 
             if (widgetPart == null)
                 return HttpNotFound();
+            if (!_services.Authorizer.Authorize(
+                Orchard.Core.Contents.Permissions.DeleteContent, widgetPart, T("You do not have permission to delete the widget.")))
+                return new HttpUnauthorizedResult();
 
             try {
                 _widgetsService.DeleteWidget(widgetPart.Id);
